+++ packages/most_schema_parser/lib/src/properties/most_property.dart
@@ -2,7 +2,7 @@ part 'most_object_property.dart';
 part 'most_value_property.dart';
 
 /// Object model JSON Schema property representation.
-abstract interface class MostProperty {
+sealed class MostProperty {
   /// Title.
   /// Mapped from `title`.
   final String? title;