+++ packages/most_schema_parser/lib/src/properties/most_value_property.dart
@@ -1,7 +1,7 @@
 part of 'most_property.dart';
 
 /// Object model JSON Schema property (without children) representation.
-class MostValueProperty extends MostProperty {
+abstract class MostValueProperty extends MostProperty {
   /// Creates an object model for the JSON Schema property (without children).
   const MostValueProperty({
     required super.title,