+++ packages/most_schema_parser/lib/src/validator/most_validator.dart
@@ -1,6 +1,0 @@
-import 'models/validation_message.dart';
-
-abstract class MostValidator {
-  /// Returns validation messages (errors and warning) or an empty list.
-  List<ValidationMessage> validate(Map<String, dynamic> map);
-}